Upgrade dependencies

+++ src/auth/signin.ts
@@ -7,6 +7,7 @@ import Purchases from 'react-native-purchases';
 import { ARGON2ID_ITERATIONS, HKDF_PASSWORD_TOKEN_LENGTH } from '@crypto';
 
 import type { SignInResponse } from '@models/SignIn';
+
 import { API_URL, sendPostRequest } from '@utils/sendRequest';
 
 import decryptProfileData from './decryptProfileData';
@@ -34,7 +35,7 @@ async function fetchSignIn(
     passwordHashSalt,
     32,
     ARGON2ID_ITERATIONS,
-    32768 * 1024,
+    BigInt(32768 * 1024),
   );
 
   const authPasswordTokenBuffer = new HKDF(
