+++ src/auth/signup.ts
@@ -6,6 +6,7 @@ import { argon2idDeriveKey } from 'react-native-nacl-jsi';
 import { ARGON2ID_ITERATIONS, HKDF_PASSWORD_TOKEN_LENGTH } from '@crypto';
 
 import type { SignUpResponse } from '@models/SignUp';
+
 import { API_URL, sendPostRequest } from '@utils/sendRequest';
 
 import encryptProfileData from './encryptProfileData';
@@ -21,7 +22,7 @@ export default async function signUp(username: string, password: string): Promis
     profileData.passwordHashSalt,
     32,
     ARGON2ID_ITERATIONS,
-    32768 * 1024,
+    BigInt(32768 * 1024),
   );
 
   const authPasswordTokenBuffer = new HKDF(
